+++ dart/geobase/lib/src/vector_data/model/geometry/multi_linestring.dart
@@ -183,7 +183,7 @@ class MultiLineString extends SimpleGeometry {
       _lineStrings.isNotEmpty ? _lineStrings.first.type : Coords.xy;
 
   @override
-  bool get isEmpty => _lineStrings.isEmpty;
+  bool get isEmptyByGeometry => _lineStrings.isEmpty;
 
   /// The chains of all line strings.
   List<PositionArray> get chains => _lineStrings;
@@ -201,7 +201,7 @@ class MultiLineString extends SimpleGeometry {
   @override
   @Deprecated('Use populated or unpopulated instead.')
   MultiLineString bounded({bool recalculate = false}) {
-    if (isEmpty) return this;
+    if (isEmptyByGeometry) return this;
 
     if (recalculate || bounds == null) {
       // return a new MultiLineString (chains kept intact) with populated bounds
@@ -225,7 +225,7 @@ class MultiLineString extends SimpleGeometry {
   }) {
     if (onBounds) {
       // create a new geometry if bounds was unpopulated and geometry not empty
-      if (bounds == null && !isEmpty) {
+      if (bounds == null && !isEmptyByGeometry) {
         return MultiLineString(
           chains,
           bounds: BoundsBuilder.calculateBounds(
@@ -272,14 +272,15 @@ class MultiLineString extends SimpleGeometry {
   }
 
   @override
-  void writeTo(SimpleGeometryContent writer, {String? name}) => isEmpty
-      ? writer.emptyGeometry(Geom.multiLineString, name: name)
-      : writer.multiLineString(
-          _lineStrings,
-          type: coordType,
-          name: name,
-          bounds: bounds,
-        );
+  void writeTo(SimpleGeometryContent writer, {String? name}) =>
+      isEmptyByGeometry
+          ? writer.emptyGeometry(Geom.multiLineString, name: name)
+          : writer.multiLineString(
+              _lineStrings,
+              type: coordType,
+              name: name,
+              bounds: bounds,
+            );
 
   // NOTE: coordinates as raw data
 
@@ -308,7 +309,7 @@ class MultiLineString extends SimpleGeometry {
   }) {
     assertTolerance(toleranceHoriz);
     if (other is! MultiLineString) return false;
-    if (isEmpty || other.isEmpty) return false;
+    if (isEmptyByGeometry || other.isEmptyByGeometry) return false;
     if (bounds != null &&
         other.bounds != null &&
         !bounds!.equals2D(
@@ -343,7 +344,7 @@ class MultiLineString extends SimpleGeometry {
     assertTolerance(toleranceHoriz);
     assertTolerance(toleranceVert);
     if (other is! MultiLineString) return false;
-    if (isEmpty || other.isEmpty) return false;
+    if (isEmptyByGeometry || other.isEmptyByGeometry) return false;
     if (!coordType.is3D || !other.coordType.is3D) return false;
     if (bounds != null &&
         other.bounds != null &&